project_assets/tests/TextSeg/requirements.txt
pandas
sentence-transformers
numpy<2
Pillow
faiss-gpu